+++ cadc-inventory-util/build.gradle
@@ -16,7 +16,7 @@ sourceCompatibility = 1.8
 
 group = 'org.opencadc'
 
-version = '0.1.8'
+version = '0.1.9'
 
 description = 'OpenCADC Storage Inventory client utility library'
 def git_url = 'https://github.com/opencadc/storage-inventory'
