+++ cadc-storage-adapter-ad/build.gradle
@@ -25,7 +25,7 @@ dependencies {
     compile 'org.opencadc:cadc-dali:[1.2.1,)'
     compile 'org.opencadc:cadc-tap:[1.0.1,)'
     compile 'org.opencadc:cadc-inventory:[0.7,)'
-    compile 'org.opencadc:cadc-storage-adapter:[0.9,)'
+    compile 'org.opencadc:cadc-storage-adapter:[0.10.1,)'
     compile 'org.opencadc:cadc-registry:[1.0,)'
     compile 'org.opencadc:cadc-vos:[1.2,2.0)'
