+++ cadc-storage-adapter-swift/build.gradle
@@ -25,7 +25,7 @@ dependencies {
     compile 'org.apache.commons:commons-pool2:[2.9,3.0)'
     compile 'org.opencadc:cadc-util:[1.6,2.0)'
     compile 'org.opencadc:cadc-inventory:[0.7,)'
-    compile 'org.opencadc:cadc-storage-adapter:[0.10,)'
+    compile 'org.opencadc:cadc-storage-adapter:[0.10.1,)'
 
     // swift API
     compile 'org.opencadc:joss:0.10.4-pdowler1'
